requirements-minimal.txt
Flask>=2.0.1,<3.0.0
Flask-SQLAlchemy>=2.5.1,<3.0.0
Flask-Login>=0.5.0,<0.7.0
Flask-WTF>=0.15.1,<1.2.0
email_validator>=1.1.3,<2.1.0
Werkzeug>=2.0.1,<3.0.0
python-dotenv>=0.19.2,<1.1.0
joblib>=1.0.1,<1.4.0
